+++ packages/api/tests/atp-agent.test.ts
@@ -283,6 +283,9 @@ describe('agent', () => {
           threadgateAllowRules: undefined,
           postgateEmbeddingRules: undefined,
         },
+        verificationPrefs: {
+          hideBadges: false,
+        },
       })
 
       await agent.setAdultContentEnabled(true)
@@ -328,6 +331,9 @@ describe('agent', () => {
         postInteractionSettings: {
           threadgateAllowRules: undefined,
           postgateEmbeddingRules: undefined,
+        },
+        verificationPrefs: {
+          hideBadges: false,
         },
       })
 
@@ -375,6 +381,9 @@ describe('agent', () => {
           threadgateAllowRules: undefined,
           postgateEmbeddingRules: undefined,
         },
+        verificationPrefs: {
+          hideBadges: false,
+        },
       })
 
       await agent.setContentLabelPref('misinfo', 'hide')
@@ -420,6 +429,9 @@ describe('agent', () => {
         postInteractionSettings: {
           threadgateAllowRules: undefined,
           postgateEmbeddingRules: undefined,
+        },
+        verificationPrefs: {
+          hideBadges: false,
         },
       })
 
@@ -470,6 +482,9 @@ describe('agent', () => {
         postInteractionSettings: {
           threadgateAllowRules: undefined,
           postgateEmbeddingRules: undefined,
+        },
+        verificationPrefs: {
+          hideBadges: false,
         },
       })
 
@@ -524,6 +539,9 @@ describe('agent', () => {
           threadgateAllowRules: undefined,
           postgateEmbeddingRules: undefined,
         },
+        verificationPrefs: {
+          hideBadges: false,
+        },
       })
 
       await agent.addPinnedFeed('at://bob.com/app.bsky.feed.generator/fake')
@@ -576,6 +594,9 @@ describe('agent', () => {
         postInteractionSettings: {
           threadgateAllowRules: undefined,
           postgateEmbeddingRules: undefined,
+        },
+        verificationPrefs: {
+          hideBadges: false,
         },
       })
 
@@ -630,6 +651,9 @@ describe('agent', () => {
           threadgateAllowRules: undefined,
           postgateEmbeddingRules: undefined,
         },
+        verificationPrefs: {
+          hideBadges: false,
+        },
       })
 
       await agent.removeSavedFeed('at://bob.com/app.bsky.feed.generator/fake')
@@ -683,6 +707,9 @@ describe('agent', () => {
           threadgateAllowRules: undefined,
           postgateEmbeddingRules: undefined,
         },
+        verificationPrefs: {
+          hideBadges: false,
+        },
       })
 
       await agent.addPinnedFeed('at://bob.com/app.bsky.feed.generator/fake')
@@ -735,6 +762,9 @@ describe('agent', () => {
         postInteractionSettings: {
           threadgateAllowRules: undefined,
           postgateEmbeddingRules: undefined,
+        },
+        verificationPrefs: {
+          hideBadges: false,
         },
       })
 
@@ -795,6 +825,9 @@ describe('agent', () => {
           threadgateAllowRules: undefined,
           postgateEmbeddingRules: undefined,
         },
+        verificationPrefs: {
+          hideBadges: false,
+        },
       })
 
       await agent.removeSavedFeed('at://bob.com/app.bsky.feed.generator/fake')
@@ -847,6 +880,9 @@ describe('agent', () => {
         postInteractionSettings: {
           threadgateAllowRules: undefined,
           postgateEmbeddingRules: undefined,
+        },
+        verificationPrefs: {
+          hideBadges: false,
         },
       })
 
@@ -901,6 +937,9 @@ describe('agent', () => {
           threadgateAllowRules: undefined,
           postgateEmbeddingRules: undefined,
         },
+        verificationPrefs: {
+          hideBadges: false,
+        },
       })
 
       await agent.setFeedViewPrefs('home', { hideReplies: true })
@@ -954,6 +993,9 @@ describe('agent', () => {
           threadgateAllowRules: undefined,
           postgateEmbeddingRules: undefined,
         },
+        verificationPrefs: {
+          hideBadges: false,
+        },
       })
 
       await agent.setFeedViewPrefs('home', { hideReplies: false })
@@ -1006,6 +1048,9 @@ describe('agent', () => {
         postInteractionSettings: {
           threadgateAllowRules: undefined,
           postgateEmbeddingRules: undefined,
+        },
+        verificationPrefs: {
+          hideBadges: false,
         },
       })
 
@@ -1067,6 +1112,9 @@ describe('agent', () => {
           threadgateAllowRules: undefined,
           postgateEmbeddingRules: undefined,
         },
+        verificationPrefs: {
+          hideBadges: false,
+        },
       })
 
       await agent.setThreadViewPrefs({ sort: 'random' })
@@ -1126,6 +1174,9 @@ describe('agent', () => {
         postInteractionSettings: {
           threadgateAllowRules: undefined,
           postgateEmbeddingRules: undefined,
+        },
+        verificationPrefs: {
+          hideBadges: false,
         },
       })
 
@@ -1187,6 +1238,9 @@ describe('agent', () => {
           threadgateAllowRules: undefined,
           postgateEmbeddingRules: undefined,
         },
+        verificationPrefs: {
+          hideBadges: false,
+        },
       })
 
       await agent.setInterestsPref({ tags: ['foo', 'bar'] })
@@ -1246,6 +1300,9 @@ describe('agent', () => {
         postInteractionSettings: {
           threadgateAllowRules: undefined,
           postgateEmbeddingRules: undefined,
+        },
+        verificationPrefs: {
+          hideBadges: false,
         },
       })
     })
@@ -1434,6 +1491,9 @@ describe('agent', () => {
           threadgateAllowRules: undefined,
           postgateEmbeddingRules: undefined,
         },
+        verificationPrefs: {
+          hideBadges: false,
+        },
       })
 
       await agent.setAdultContentEnabled(false)
@@ -1495,6 +1555,9 @@ describe('agent', () => {
         postInteractionSettings: {
           threadgateAllowRules: undefined,
           postgateEmbeddingRules: undefined,
+        },
+        verificationPrefs: {
+          hideBadges: false,
         },
       })
 
@@ -1559,6 +1622,9 @@ describe('agent', () => {
           threadgateAllowRules: undefined,
           postgateEmbeddingRules: undefined,
         },
+        verificationPrefs: {
+          hideBadges: false,
+        },
       })
 
       await agent.removeLabeler('did:plc:other')
@@ -1617,6 +1683,9 @@ describe('agent', () => {
         postInteractionSettings: {
           threadgateAllowRules: undefined,
           postgateEmbeddingRules: undefined,
+        },
+        verificationPrefs: {
+          hideBadges: false,
         },
       })
 
@@ -1677,6 +1746,9 @@ describe('agent', () => {
           threadgateAllowRules: undefined,
           postgateEmbeddingRules: undefined,
         },
+        verificationPrefs: {
+          hideBadges: false,
+        },
       })
 
       await agent.setPersonalDetails({ birthDate: '2023-09-11T18:05:42.556Z' })
@@ -1735,6 +1807,9 @@ describe('agent', () => {
         postInteractionSettings: {
           threadgateAllowRules: undefined,
           postgateEmbeddingRules: undefined,
+        },
+        verificationPrefs: {
+          hideBadges: false,
         },
       })
 
@@ -1806,6 +1881,9 @@ describe('agent', () => {
         postInteractionSettings: {
           threadgateAllowRules: undefined,
           postgateEmbeddingRules: undefined,
+        },
+        verificationPrefs: {
+          hideBadges: false,
         },
       })
 
@@ -3501,6 +3579,37 @@ describe('agent', () => {
             postgateEmbeddingRules: [],
           }),
         ).rejects.toThrow()
+      })
+    })
+
+    describe('setVerificationPrefs', () => {
+      let agent: AtpAgent
+
+      beforeAll(async () => {
+        agent = new AtpAgent({ service: network.pds.url })
+
+        await agent.createAccount({
+          handle: 'verification-prefs.test',
+          email: 'verification-prefs@test.com',
+          password: 'password',
+        })
+      })
+
+      it('default state', async () => {
+        const next: AppBskyActorDefs.VerificationPrefs = {
+          hideBadges: false,
+        }
+        const prefs = await agent.getPreferences()
+        expect(prefs.verificationPrefs).toEqual(next)
+      })
+
+      it('updates', async () => {
+        const next: AppBskyActorDefs.VerificationPrefs = {
+          hideBadges: true,
+        }
+        await agent.setVerificationPrefs(next)
+        const prefs = await agent.getPreferences()
+        expect(prefs.verificationPrefs).toEqual(next)
       })
     })