+++ packages/bsky/src/hydration/actor.ts
@@ -1,6 +1,8 @@
+import { mapDefined } from '@atproto/common'
 import { DataPlaneClient } from '../data-plane/client'
 import { Record as ProfileRecord } from '../lexicon/types/app/bsky/actor/profile'
 import { Record as ChatDeclarationRecord } from '../lexicon/types/chat/bsky/actor/declaration'
+import { VerificationMeta } from '../proto/bsky_pb'
 import {
   HydrationMap,
   RecordInfo,
@@ -23,7 +25,19 @@ export type Actor = {
   upstreamStatus?: string
   createdAt?: Date
   priorityNotifications: boolean
+  trustedVerifier?: boolean
+  verifications: VerificationHydrationState[]
 }
+
+export type VerificationHydrationState = {
+  issuer: string
+  uri: string
+  handle: string
+  displayName: string
+  createdAt: string
+}
+
+export type VerificationMetaRequired = Required<VerificationMeta>
 
 export type Actors = HydrationMap<Actor>
 
@@ -127,6 +141,28 @@ export class ActorHydrator {
         ? parseRecord<ProfileRecord>(actor.profile, includeTakedowns)
         : undefined
 
+      const verifications = mapDefined(
+        Object.entries(actor.verifiedBy),
+        ([actorDid, verificationMeta]) => {
+          if (
+            verificationMeta.displayName &&
+            verificationMeta.handle &&
+            verificationMeta.rkey &&
+            verificationMeta.sortedAt
+          ) {
+            return {
+              issuer: actorDid,
+              uri: `at://${actorDid}/app.bsky.graph.verification/${verificationMeta.rkey}`,
+              handle: verificationMeta.handle,
+              displayName: verificationMeta.displayName,
+              createdAt: verificationMeta.sortedAt.toDate().toISOString(),
+            }
+          }
+          // Filter out the verification meta that doesn't contain all info.
+          return undefined
+        },
+      )
+
       return acc.set(did, {
         did,
         handle: parseString(actor.handle),
@@ -141,6 +177,8 @@ export class ActorHydrator {
         upstreamStatus: actor.upstreamStatus || undefined,
         createdAt: actor.createdAt?.toDate(),
         priorityNotifications: actor.priorityNotifications,
+        trustedVerifier: actor.trustedVerifier,
+        verifications,
       })
     }, new HydrationMap<Actor>())
   }