+++ packages/dev-env/src/seed/client.ts
@@ -7,6 +7,7 @@ import {
   AppBskyGraphBlock,
   AppBskyGraphFollow,
   AppBskyGraphList,
+  AppBskyGraphVerification,
   AppBskyRichtextFacet,
   AtpAgent,
   ComAtprotoModerationCreateReport,
@@ -113,6 +114,9 @@ export class SeedClient<
       }
     >
   >
+
+  verifications: Record<string, Record<string, AtUri>>
+
   dids: Record<string, string>
 
   constructor(
@@ -130,6 +134,7 @@ export class SeedClient<
     this.lists = {}
     this.feedgens = {}
     this.starterpacks = {}
+    this.verifications = {}
     this.dids = {}
   }
 
@@ -553,6 +558,29 @@ export class SeedClient<
       },
     )
     return result.data
+  }
+
+  async verify(
+    by: string,
+    subject: string,
+    handle: string,
+    displayName: string,
+    overrides?: Partial<AppBskyGraphVerification.Record>,
+  ) {
+    const res = await this.agent.app.bsky.graph.verification.create(
+      { repo: by },
+      {
+        subject,
+        createdAt: new Date().toISOString(),
+        handle,
+        displayName,
+        ...overrides,
+      },
+      this.getHeaders(by),
+    )
+    this.verifications[by] ??= {}
+    this.verifications[by][subject] = new AtUri(res.uri)
+    return this.verifications[by][subject]
   }
 
   getHeaders(did: string) {